+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/design/ExcessiveParameterListRule.java
@@ -38,11 +38,11 @@ public class ExcessiveParameterListRule extends AbstractJavaCounterCheckRule<AST
     }
 
     /**
-     * @deprecated since 7.18.0
+     * @deprecated since 7.18.0. This method is not used anymore and shouldn't be implemented.
      */
-    @Override
+    @Deprecated
     protected boolean isViolation(ASTFormalParameters node, int reportLevel) {
-        return super.isViolation(node, reportLevel);
+        throw new UnsupportedOperationException("method is deprecated and not supported anymore.");
     }
 
     @Override