+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/design/ExcessivePublicCountRule.java
@@ -40,11 +40,11 @@ public class ExcessivePublicCountRule extends AbstractJavaCounterCheckRule<ASTTy
     }
 
     /**
-     * @deprecated since 7.18.0
+     * @deprecated since 7.18.0. This method is not used anymore and shouldn't be implemented.
      */
-    @Override
+    @Deprecated
     protected boolean isViolation(ASTTypeDeclaration node, int reportLevel) {
-        return super.isViolation(node, reportLevel);
+        throw new UnsupportedOperationException("method is deprecated and not supported anymore.");
     }
 
     @Override